小错

+++ hospital-back-end/src/main/java/com/neuedu/hospitalbackend/service/serviceimplementation/doctorstationservice/PreliminaryCaseServiceImpl.java
@@ -162,11 +162,9 @@ public class PreliminaryCaseServiceImpl implements PreliminaryCaseService {
         List<String> existedDiseaseIcdCodes = diagnoseMapper.listDiseaseIcdCodesByCaseId(caseId); //所有数据库暂存诊断
         if(diagnoseType == 0) {
             diagnoses = patientCaseParam.getTraditionalDiagnose();
-            System.out.println("0000000000000");
         }
         else if(diagnoseType == 1) {
             diagnoses = patientCaseParam.getModernDiagnose();
-            System.out.println("11111111111111111");
         }
         else
             return CommonResult.fail(ResultCode.E_801);
@@ -193,7 +191,7 @@ public class PreliminaryCaseServiceImpl implements PreliminaryCaseService {
             }
         }
         //若数据库已存该诊断，但不再暂存/提交，则删除该诊断
-        while(!existedDiseaseIcdCodes.isEmpty()) {
+        if(!existedDiseaseIcdCodes.isEmpty()) {
             for (String leftDiseaseIcdCode : existedDiseaseIcdCodes) {
                 count = diagnoseMapper.deleteByCaseIdAndDiseaseIcdCode(caseId, leftDiseaseIcdCode);
             }
